// src/types/lineas.ts
export interface LineaRow {
  tarifa: string
  perfil: string
  observaciones: string
  numero: string
  extension: string
  usuario: string
  imei: string
  pin: string
  sim2: string
  pukSim2: string
  powerAppsId: string
  _rowIndex: number
  _sheet: string
}

export interface TarifaInfo {
  nombre: string
  datos: string
}